requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.9
langchain-openai>=0.0.3
langgraph>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.2.0
openai>=1.3.8
python-dotenv>=1.0.0
pandas>=2.1.4
numpy>=1.24.3
tiktoken>=0.5.2
pydantic>=2.5.2 